Move consuming sectors back into RootSectorNode
